refactor(bridge): move position_query tests to separate file; name cartesian tag
